- Update drink's favorite field

+++ app/src/main/java/cl/gerardomascayano/drinksapp/data/DrinkDataSource.kt
@@ -14,6 +14,6 @@ interface DrinkDataSource {
 
     suspend fun getDrinksByName(name: String): List<DrinkSearchTuple>
 
-    suspend fun updateDrink(drinkEntity: DrinkEntity): Long
+    suspend fun updateDrink(drinkEntity: DrinkEntity): Int
 
 }
