+++ app/src/main/java/cl/gerardomascayano/drinksapp/data/LocalDrinkDataSourceImpl.kt
@@ -12,7 +12,7 @@ class LocalDrinkDataSourceImpl @Inject constructor(private val drinksDao: DrinkD
     override suspend fun getAllFavoriteDrinks(): List<DrinkWithIngredients> = drinksDao.getDrinkWithIngredientsFavorite()
     override suspend fun getDrinkById(drinkId: Int): DrinkWithIngredients = drinksDao.getDrinkById(drinkId)
     override suspend fun getDrinksByName(name: String): List<DrinkSearchTuple> = drinksDao.getDrinksByName(name)
-    override suspend fun updateDrink(drinkEntity: DrinkEntity): Long = drinksDao.updateDrink(drinkEntity)
+    override suspend fun updateDrink(drinkEntity: DrinkEntity): Int = drinksDao.updateDrink(drinkEntity)
 
     override suspend fun getAllUnfavoriteDrinks(): List<DrinkWithIngredients> = drinksDao.getDrinkWithIngredientsUnfavorite()
 }
