+++ app/src/main/java/cl/gerardomascayano/drinksapp/data/db/entities/drink/DrinkDao.kt
@@ -24,7 +24,7 @@ interface DrinkDao {
     suspend fun getDrinkWithIngredientsUnfavorite(): List<DrinkWithIngredients>
 
     @Update(entity = DrinkEntity::class)
-    suspend fun updateDrink(drinkEntity: DrinkEntity): Long
+    suspend fun updateDrink(drinkEntity: DrinkEntity): Int
 
 
 }
